Add improved memory deallocation handling (#43)
* Add improved memory deallocation handling

This commit addresses some issues seen when running the Device Defender code shortly after machine startup, which can result in a message /net/proc/tcp file. When there are invalid lines in the file, this can trigger early memory deallocation of structs such as aws_string that have not actually been allocated. This commit resolves this issue by adding additional validation to ensure we're only deallocating memory that has actually been utilized, and also handling to ensure we don't perform a double free.

* Use AWS_ZERO_STRUCT instead of boolean flags

This commit introduces the use of AWS_ZERO_STRUCT to nullify members of the connection rather than using boolean flags to detect if regions of the connection have not had memory allocated. This does not fix any issues, it is simply a stylistic change with less condition checks.

* Cleanup import and usage of AWS_ZERO_STRUCT

This commit removes and unused import statement and consolidates the usage of AWS_ZERO_STRUCT into a single line.

Co-authored-by: Matt <mwp@amazon.com>